Fix compiler warning

+++ SourceX/DiabloUI/credits.cpp
@@ -117,7 +117,7 @@ public:
 		ArtBackground.Unload();
 		UnloadTtfFont();
 
-		for (int x = 0; x < lines_.size(); x++) {
+		for (size_t x = 0; x < lines_.size(); x++) {
 			if (lines_[x].m_surface)
 				SDL_FreeSurface(lines_[x].m_surface);
 		}
